refactor: update PostCSS and Tailwind configuration
- Changed PostCSS configuration to use '@tailwindcss/postcss' plugin.
- Updated global CSS to import Tailwind CSS instead of using @tailwind directives.
- Refactored imports in the Home page to use the new Card component path.
- Added new Button and Card components in the common UI directory.
- Updated Links component to import the Button from the new common path.
- Removed the old Tailwind configuration file as it is no longer needed.
