+++ web/package.json
@@ -30,8 +30,7 @@
     "eslint-config-next": "15.1.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-react-compiler": "19.1.0-rc.2",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
+    "tailwindcss": "^4.1.11",
     "typescript": "^5"
   }
 }
